+++ config/routes.rb
@@ -1,12 +1,14 @@
 Rails.application.routes.draw do
   ActiveAdmin.routes(self)
-  devise_for :users, controllers: {confirmations: 'confirmations', registrations: 'registrations', sessions: 'users/sessions'}
+
+  devise_for :users, controllers: { sessions:           'users/sessions',
+                                    omniauth_callbacks: 'users/omniauth_callbacks' }
 
   devise_scope :user do
-    patch "/confirm" => "confirmations#confirm"
+    patch '/confirm' => 'confirmations#confirm'
   end
 
-  get '/switch_chapter', to: "application#switch_chapter", as: :switch_chapter
+  get '/switch_chapter', to: 'application#switch_chapter', as: :switch_chapter
 
   resources :characters, except: [:edit, :destroy]
   resources :projects